api/pyproject.toml
[project]
name = "fintech-payment-api"
version = "0.1.0"
description = "Production-grade payment API infrastructure with security compliance and multi-environment CI/CD"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.2",
    "fastapi[standard]>=0.128.0",
    "psycopg>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
